chore(vscode): add tests for duplicate model and bad model block
